# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ResearchSphere repository API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  STUDENT: 'STUDENT',
  REVIEWER: 'REVIEWER',
  ADMIN: 'ADMIN',
} as const;

export interface UserProfile {
  id: string;
  name: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: UserRole;
  /** @nullable */
  departmentId: string | null;
  /** @nullable */
  departmentName: string | null;
}

export type PaperStatus = typeof PaperStatus[keyof typeof PaperStatus];


export const PaperStatus = {
  PENDING_REVIEW: 'PENDING_REVIEW',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  REVISION_REQUIRED: 'REVISION_REQUIRED',
} as const;

export type ReviewDecision = typeof ReviewDecision[keyof typeof ReviewDecision];


export const ReviewDecision = {
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  REVISION_REQUIRED: 'REVISION_REQUIRED',
} as const;

export type CategoryKind = typeof CategoryKind[keyof typeof CategoryKind];


export const CategoryKind = {
  RESEARCH_AREA: 'RESEARCH_AREA',
  PAPER_TYPE: 'PAPER_TYPE',
} as const;

export interface PaperVersion {
  versionNumber: number;
  /** @nullable */
  objectPath: string | null;
  createdAt: string;
}

export interface PaperComment {
  id: string;
  userId: string;
  userName: string;
  content: string;
  createdAt: string;
}

export interface Paper {
  id: string;
  title: string;
  abstract: string;
  year: number;
  departmentId: string;
  departmentName: string;
  researchArea: string;
  paperType: string;
  /** @nullable */
  doi: string | null;
  /** @nullable */
  objectPath: string | null;
  /** @nullable */
  fileHash: string | null;
  keywords: string[];
  status: PaperStatus;
  uploadedById: string;
  authorName: string;
  createdAt: string;
  versionNumber: number;
  /** @nullable */
  latestReviewComment: string | null;
  versions: PaperVersion[];
  comments: PaperComment[];
}

export interface CommentInput {
  /** @minLength 1 */
  content: string;
}

export interface Collection {
  id: string;
  name: string;
  paperCount: number;
  createdAt: string;
  papers?: Paper[];
}

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
}

export interface PaperDetail {
  paper: Paper;
  related: Paper[];
}

export interface PaperInput {
  /** @minLength 1 */
  title: string;
  abstract: string;
  year: number;
  departmentId: string;
  researchArea: string;
  paperType: string;
  /** @nullable */
  doi?: string | null;
  /** @minLength 1 */
  objectPath: string;
  keywords: string[];
}

export interface PaperRevisionInput {
  /** @minLength 1 */
  title: string;
  abstract: string;
  /** @minLength 1 */
  objectPath: string;
  keywords: string[];
}

export interface MetadataPreviewInput {
  /** @minLength 1 */
  objectPath: string;
}

export interface MetadataPreview {
  title: string;
  abstract: string;
  keywords: string[];
  doi?: string | null;
  fileHash: string;
  duplicate: boolean;
}

export interface ReviewInput {
  status: ReviewDecision;
  /** @minLength 1 */
  comments: string;
}

export type MyAnalyticsPapersItem = {
  id: string;
  title: string;
  views: number;
  downloads: number;
  bookmarks: number;
};

export interface MyAnalytics {
  totalViews: number;
  totalDownloads: number;
  totalBookmarks: number;
  papers: MyAnalyticsPapersItem[];
}

export interface BookmarkState {
  bookmarked: boolean;
}

export interface Department {
  id: string;
  name: string;
  code: string;
}

export interface DepartmentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 2 */
  code: string;
}

export interface Category {
  id: string;
  kind: CategoryKind;
  name: string;
}

export interface CategoryInput {
  kind: CategoryKind;
  /** @minLength 1 */
  name: string;
}

export interface UserRoleInput {
  role: UserRole;
  /** @nullable */
  departmentId?: string | null;
}

export interface AdminUser {
  id: string;
  name: string;
  /** @nullable */
  email: string | null;
  role: UserRole;
  /** @nullable */
  departmentId: string | null;
  /** @nullable */
  departmentName: string | null;
}

export interface AnalyticsGroup {
  label: string;
  count: number;
}

export interface PaperActivityMetric {
  id: string;
  title: string;
  count: number;
}

export interface AdminAnalytics {
  totalPapers: number;
  totalUsers: number;
  departmentsCount: number;
  totalViewsDownloads: number;
  papersByDepartment: AnalyticsGroup[];
  papersByYear: AnalyticsGroup[];
  submissionsByStatus: AnalyticsGroup[];
  mostViewedPapers: PaperActivityMetric[];
  mostDownloadedPapers: PaperActivityMetric[];
}

export type UploadUrlInputContentType = typeof UploadUrlInputContentType[keyof typeof UploadUrlInputContentType];


export const UploadUrlInputContentType = {
  'application/pdf': 'application/pdf',
} as const;

export interface UploadUrlInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  contentType: UploadUrlInputContentType;
}

export type UploadUrlResponseMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlResponseMetadata;
}

export interface LoginUserBody {
  email: string;
  password: string;
}

export interface RegisterUserBody {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface CreateAdminUserInput {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  /** @nullable */
  departmentId?: string | null;
}

/**
 * Invalid input.
 */
export type BadRequestResponse = ErrorEnvelope;

/**
 * Sign in is required.
 */
export type UnauthorizedResponse = ErrorEnvelope;

/**
 * The current user does not have permission.
 */
export type ForbiddenResponse = ErrorEnvelope;

/**
 * Resource not found.
 */
export type NotFoundResponse = ErrorEnvelope;

/**
 * The server could not complete the request.
 */
export type ServerErrorResponse = ErrorEnvelope;

export type AuthorizationSessionHeaderParameter = string;

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type ListPapersParams = {
q?: string;
semantic?: boolean;
departmentId?: string;
year?: number;
researchArea?: string;
paperType?: string;
};

export type ListCategoriesParams = {
kind?: CategoryKind;
};