# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ResearchSphere repository API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [Administration]
      summary: Get system-wide research analytics
      responses:
        "200":
          description: Summary metrics and grouped breakdowns.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnalytics"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/departments:
    post:
      operationId: createDepartment
      tags: [Administration]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Department created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/departments/{departmentId}:
    patch:
      operationId: updateDepartment
      tags: [Administration]
      summary: Update a department
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated department.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDepartment
      tags: [Administration]
      summary: Delete an unused department
      parameters:
        - $ref: "#/components/parameters/DepartmentId"
      responses:
        "204":
          description: Department deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Department is referenced by users or papers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/categories:
    get:
      operationId: listCategories
      tags: [Administration]
      summary: List research areas or paper types
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CategoryKind"
      responses:
        "200":
          description: Available categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [Administration]
      summary: Add a research area or paper type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/categories/{categoryId}:
    patch:
      operationId: updateCategory
      tags: [Administration]
      summary: Rename a research area or paper type
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCategory
      tags: [Administration]
      summary: Delete an unused category
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Category is referenced by papers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [Administration]
      summary: List users for role and reviewer management
      responses:
        "200":
          description: App users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createAdminUser
      tags: [Administration]
      summary: Create a new user with a specific role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserInput"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [Administration]
      summary: Assign a user role and department
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleInput"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    DepartmentId:
      name: departmentId
      in: path
      required: true
      schema:
        type: string
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Sign in is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The current user does not have permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UserRole:
      type: string
      enum: [STUDENT, REVIEWER, ADMIN]
...
    CategoryKind:
      type: string
      enum: [RESEARCH_AREA, PAPER_TYPE]
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
      required: [id, name, code]
    DepartmentInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 2
      required: [name, code]
    Category:
      type: object
      properties:
        id:
          type: string
        kind:
          $ref: "#/components/schemas/CategoryKind"
        name:
          type: string
      required: [id, kind, name]
    CategoryInput:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/CategoryKind"
        name:
          type: string
          minLength: 1
      required: [kind, name]
    UserRoleInput:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
        departmentId:
          type: [string, "null"]
      required: [role]
    AdminUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: [string, "null"]
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        departmentId:
          type: [string, "null"]
        departmentName:
          type: [string, "null"]
      required: [id, name, email, role, departmentId, departmentName]
    AnalyticsGroup:
      type: object
      properties:
        label:
          type: string
        count:
          type: integer
      required: [label, count]
    AdminAnalytics:
      type: object
      properties:
        totalPapers:
          type: integer
        totalUsers:
          type: integer
        departmentsCount:
          type: integer
        totalViewsDownloads:
          type: integer
        papersByDepartment:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsGroup"
        papersByYear:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsGroup"
        submissionsByStatus:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsGroup"
        mostViewedPapers:
          type: array
          items:
            $ref: "#/components/schemas/PaperActivityMetric"
        mostDownloadedPapers:
          type: array
          items:
            $ref: "#/components/schemas/PaperActivityMetric"
      required: [totalPapers, totalUsers, departmentsCount, totalViewsDownloads, papersByDepartment, papersByYear, submissionsByStatus, mostViewedPapers, mostDownloadedPapers]
    PaperActivityMetric:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        count:
          type: integer
      required: [id, title, count]
...
    CreateAdminUserInput:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email: { type: string }
        password: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        departmentId:
          type: [string, "null"]